% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpm_gamma_mixture_single_scale.R
\name{ebpm_gamma_mixture_single_scale}
\alias{ebpm_gamma_mixture_single_scale}
\title{Empirical Bayes Poisson Mean with Mixture of Gamma as Prior (choosing one scale and varying shape)}
\usage{
ebpm_gamma_mixture_single_scale(x, s = 1, scale = "estimate",
  theta = "max", g_init = NULL, fix_g = FALSE, m = 2,
  control = NULL, low = NULL)
}
\arguments{
\item{x}{A vector of Poisson observations.}

\item{s}{A vector of scaling factors for Poisson observations: the model is \eqn{y[j]~Pois(s[j]*lambda[j])}.}

\item{scale}{Either \code{"estimate"} if the scale parameters are to be estimated
from the data, or A list of  \code{shape} (set to all 1s for exponential mixture) and  \code{scale} that specifies the  components of the mixture, where each pair of \code{shape}, \code{scale} described a \code{gamma(shape, scale)}.}

\item{g_init}{The prior distribution \eqn{g}, of the class \code{gammamix}. Usually this is left
unspecified (\code{NULL}) and estimated from the data. However, it can be
used in conjuction with \code{fix_g = TRUE} to fix the prior (useful, for
example, to do computations with the "true" \eqn{g} in simulations). If
\code{g_init} is specified but \code{fix_g = FALSE}, \code{g_init}
specifies the initial value of \eqn{g} used during optimization.}

\item{fix_g}{If \code{TRUE}, fix the prior \eqn{g} at \code{g_init} instead
of estimating it.}

\item{m}{multiple coefficient when selectig grid, so the  b_k is of the form {low*m^{k-1}}; must be greater than 1; default is 2}

\item{control}{A list of control parameters  to be passed to the optimization function. `mixsqp` is  used.}
}
\value{
A list containing elements:
    \describe{
      \item{\code{posterior}}{A data frame of summary results (posterior
        means, and to add posterior log mean).}
      \item{\code{fitted_g}}{The fitted prior \eqn{\hat{g}}, of class \code{gammamix}} 
      \item{\code{log_likelihood}}{The optimal log likelihood attained
        \eqn{L(\hat{g})}.}
     }
}
\description{
Uses Empirical Bayes to fit the model \deqn{x_j | \lambda_j ~ Poi(s_j \lambda_j)} with \deqn{lambda_j ~ g()}
with Mixture of Exponential: \deqn{g()  = sum_k pi_k gamma(shape = 1, rate = b_k)} 
b_k is selected to cover the lambda_i  of interest for all data  points x_i
}
\details{
The model is fit in 2 stages: i) estimate \eqn{g} by maximum likelihood (over pi_k)
ii) Compute posterior distributions for \eqn{\lambda_j} given \eqn{x_j,\hat{g}}.
}
\examples{
   beta = c(rep(0,50),rexp(50))
   x = rpois(100,beta) # simulate Poisson observations
   s = replicate(100,1)
   m = 2
   out = ebpm::ebpm_gamma_mixture(x,s)
   
}
