% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpm_gh_gamma.R
\name{ebpm_gh_gamma}
\alias{ebpm_gh_gamma}
\title{Empirical Bayes Poisson Mean (gh-gamma as prior)}
\usage{
ebpm_gh_gamma(
  x,
  s = 1,
  g_init = NULL,
  fix_g = c(F, F, F, F, F),
  control = NULL
)
}
\description{
Uses Empirical Bayes to fit the model \deqn{x_j | \lambda_j ~ Poi(s_j \lambda_j)} with \deqn{s_j lambda_j ~ g()}
with g() being gh-gamma
}
