% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpm_invBeta_gamma.R
\name{ebpm_invBeta_gamma}
\alias{ebpm_invBeta_gamma}
\title{Empirical Bayes Poisson Mean (invBeta-gamma as prior)}
\usage{
ebpm_invBeta_gamma(x, s = 1, g_init = NULL, fix_g = c(F, F, F), control = NULL)
}
\description{
Uses Empirical Bayes to fit the model \deqn{x_j | \lambda_j ~ Poi(s_j \lambda_j)} with \deqn{s_j lambda_j ~ g()}
with g() being invBeta-gamma
}
