% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpm_point_gamma.R
\name{ebpm_point_gamma}
\alias{ebpm_point_gamma}
\title{Empirical Bayes Poisson Mean with Point Gamma  as Prior}
\usage{
ebpm_point_gamma(x, s, init_par = c(0.5, 1, 1), seed = 123)
}
\arguments{
\item{x}{vector of Poisson observations.}

\item{s}{vector of scale factors for Poisson observations: the model is \eqn{y[j]~Pois(scale[j]*lambda[j])}.}

\item{init_par}{vector of initialization for c(pi, a, b); default is c(0.5,1,1)}

\item{seed}{set seed (not necessary now)}
}
\description{
Uses Empirical Bayes to fit the model \deqn{x_j | \lambda_j ~ Poi(s_j \lambda_j)} with \deqn{lambda_j ~ g()}
with Point Gamma: g()  = pi_0 \delta() + (1-\pi_0) gamma(a, b)
#' @import stats
}
\details{
The model is fit in two stages: i) estimate \eqn{g} by maximum likelihood (over pi_0, a, b)
ii) Compute posterior distributions for \eqn{\lambda_j} given \eqn{x_j,\hat{g}}.
}
\examples{
   beta = c(rep(0,50),rexp(50))
   x = rpois(100,beta) # simulate Poisson observations
   s = replicate(100,1)
   out = ebpm_point_gamma(x,s)
}
