% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpm_two_gamma_fast3.R
\name{ebpm_two_gamma_fast3}
\alias{ebpm_two_gamma_fast3}
\title{Empirical Bayes Poisson Mean (mixture of two gammas as prior, faster version 3)}
\usage{
ebpm_two_gamma_fast3(
  x,
  s = 1,
  g_init = NULL,
  fix_g = FALSE,
  n_iter = 100,
  verbose = FALSE,
  control = list(tol_in = 1e-04),
  get_progress = TRUE,
  seed = 123
)
}
\arguments{
\item{x}{vector of Poisson observations.}

\item{s}{vector of scale factors for Poisson observations: the model is \eqn{y[j]~Pois(scale[j]*lambda[j])}.}

\item{g_init}{The prior distribution \eqn{g}, of the class \code{two_gamma}. Usually this is left
unspecified (\code{NULL}) and estimated from the data. However, it can be
used in conjuction with \code{fix_g = TRUE} to fix the prior (useful, for
example, to do computations with the "true" \eqn{g} in simulations). If
\code{g_init} is specified but \code{fix_g = FALSE}, \code{g_init}
specifies the initial value of \eqn{g} used during optimization.}

\item{fix_g}{If \code{TRUE}, fix the prior \eqn{g} at \code{g_init} instead
of estimating it.}

\item{control}{A list of control parameters to be passed to `mle_two_gamma` function.}

\item{get_progress}{record log_likelihood if set to TRUE}

\item{seed}{seed set to initialization if \code{g_init = NULL}}

\item{n_iter:}{number of maximum EM steps}

\item{rel_tol:}{tolerance for (maximum) relative change in parameters in \eqn{g}}
}
\value{
A list containing elements:
    \describe{
      \item{\code{posterior}}{A data frame of summary results (posterior
        means, posterior log mean).}
      \item{\code{fitted_g}}{The fitted prior \eqn{\hat{g}} of class \code{point_gamma}} 
      \item{\code{log_likelihood}}{The optimal log likelihood attained
        \eqn{L(\hat{g})}.}
     }
}
\description{
Uses Empirical Bayes to fit the model \deqn{x_j | \lambda_j ~ Poi(s_j \lambda_j)} with \deqn{lambda_j ~ g()}
with Point Gamma: g()  = pi_0 gamma(shape1, scale1) + (1-pi_0) gamma(shape2, scale2)
}
\details{
The model is fit in two stages: i) estimate \eqn{g} by maximum likelihood (over pi_0, shape, scale)
ii) Compute posterior distributions for \eqn{\lambda_j} given \eqn{x_j,\hat{g}}.
}
