% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpm_exponential_mixture.R
\name{ebpm_exponential_mixture}
\alias{ebpm_exponential_mixture}
\title{Empirical Bayes Poisson Mean with Mixture of Exponential as Prior}
\usage{
ebpm_exponential_mixture(x, s, m = 2, grid = NULL, seed = 123)
}
\description{
Empirical Bayes Poisson Mean with Mixture of Exponential as Prior
}
\section{Functions}{
\itemize{
\item \code{ebpm_exponential_mixture}: mixture weight and posterior mean
}}

