% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammamix.R
\name{gammamix}
\alias{gammamix}
\title{\code{gammamix} class}
\usage{
gammamix(pi, shape, scale)
}
\arguments{
\item{pi}{A vector of mixture proportions.}

\item{shape}{A vector of shapes}

\item{scale}{A vector of scales}
}
\description{
gammamix class that describes \deqn{g()  = sum_k pi_k gamma(shape, scale)}
}
