% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point1_gamma.R
\name{point1_gamma}
\alias{point1_gamma}
\title{\code{point1_gamma} class}
\usage{
point1_gamma(pi0, shape, scale)
}
\arguments{
\item{pi0}{A scalar specifying the weight on point mass at 1}

\item{shape}{A scalar specifying the shape of gamma}

\item{scale}{A scalar specifying the scale of gamma}
}
\description{
gammamix class that describes \deqn{g()  = pi0 delta1() + (1-pi0) gamma(shape, scale)}
}
