% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_gamma.R
\name{point_gamma}
\alias{point_gamma}
\title{\code{point_gamma} class}
\usage{
point_gamma(pi0, shape, scale)
}
\arguments{
\item{pi0}{A scalar specifying the weight on point mass at 0}

\item{shape}{A scalar specifying the shape of gamma}

\item{scale}{A scalar specifying the scale of gamma}
}
\description{
gammamix class that describes \deqn{g()  = pi0 delta() + (1-pi0) gamma(shape, scale)}
}
