% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_pois_expmix}
\alias{simulate_pois_expmix}
\title{Function for simulating from a mkixture of exponential}
\usage{
simulate_pois_expmix(n, seed = 123)
}
\description{
Function for simulating from a mkixture of exponential
}
