% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_gamma.R
\name{two_gamma}
\alias{two_gamma}
\title{\code{two_gamma} class}
\usage{
two_gamma(pi0, shape1, scale1, shape2, scale2)
}
\arguments{
\item{pi0}{A scalar specifying the weight on point mass at 0}

\item{shape1, }{A scalar specifying the shape of gamma in first component}

\item{scale1, }{A scalar specifying the scale of gamma in first component}

\item{shape2, }{A scalar specifying the shape of gamma in second component}

\item{scale2, }{A scalar specifying the scale of gamma in second component}
}
\description{
two_gamma class that describes \deqn{g()  = pi0 gamma(shape1, scale1) + (1-pi0) gamma(shape2, scale2)}
}
