% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpm_exponential_mixture.R
\name{ebpm_exponential_mixture}
\alias{ebpm_exponential_mixture}
\title{Empirical Bayes Poisson Mean with Point Gamma  as Prior}
\usage{
ebpm_exponential_mixture(x, s, m = 2, d = NULL, grid = NULL,
  seed = 123)
}
\arguments{
\item{x}{vector of Poisson observations.}

\item{s}{vector of scale factors for Poisson observations: the model is \eqn{y[j]~Pois(scale[j]*lambda[j])}.}

\item{m}{multiple coefficient when selectig grid, so the  b_k is of the form {low*m^{k-1}}; must be greater than 1; default is 2}

\item{d}{if m is not sepcified, d is number of points in the grid; if both m and d are specified, only m will be used}

\item{grid}{locations of b_k; if left NULL the algorithm automatically selects them based on data}

\item{seed}{random seed}
}
\value{
A list containing elements:
    \describe{
      \item{\code{posterior}}{A data frame of summary results (posterior
        means, and to add posterior log mean).}
      \item{\code{fitted_g}}{The fitted prior \eqn{\hat{g}}} 
      \item{\code{log_likelihood}}{The optimal log likelihood attained
        \eqn{L(\hat{g})}.}
      \item{\code{posterior_sampler}}{(TO ADD!!!) A function that can be used to
        produce samples from the posterior. It takes a single parameter
        \code{nsamp}, the number of posterior samples to return per
        observation.}
     }
}
\description{
Uses Empirical Bayes to fit the model \deqn{x_j | \lambda_j ~ Poi(s_j \lambda_j)} with \deqn{lambda_j ~ g()}
with Mixture of Exponential: g()  = sum_k pi_k gamma(1, b_k) 
b_k is selected to cover the lambda_i  of interest for all data  points x_i
}
\details{
The model is fit in 2 stages: i) estimate \eqn{g} by maximum likelihood (over pi_k)
ii) Compute posterior distributions for \eqn{\lambda_j} given \eqn{x_j,\hat{g}}.
}
\examples{
   beta = c(rep(0,50),rexp(50))
   x = rpois(100,beta) # simulate Poisson observations
   s = replicate(100,1)
   m = 2
   out = ebpm::ebpm_exponential_mixture(x,s, m)
   
}
