% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebpm_point_gamma.R
\name{ebpm_point_gamma}
\alias{ebpm_point_gamma}
\title{Empirical Bayes Poisson Mean with Point Gamma  as Prior}
\usage{
ebpm_point_gamma(x, s = 1, g_init = NULL, fix_g = F,
  pi0 = "estimate", control = NULL)
}
\arguments{
\item{x}{vector of Poisson observations.}

\item{s}{vector of scale factors for Poisson observations: the model is \eqn{y[j]~Pois(scale[j]*lambda[j])}.}

\item{g_init}{The prior distribution \eqn{g}, of the class \code{point_gamma}. Usually this is left
unspecified (\code{NULL}) and estimated from the data. However, it can be
used in conjuction with \code{fix_g = TRUE} to fix the prior (useful, for
example, to do computations with the "true" \eqn{g} in simulations). If
\code{g_init} is specified but \code{fix_g = FALSE}, \code{g_init}
specifies the initial value of \eqn{g} used during optimization.}

\item{fix_g}{If \code{TRUE}, fix the prior \eqn{g} at \code{g_init} instead
of estimating it.}

\item{pi0}{Either  \code{"estimate"} which optimizes over pi0 along with  shape and scale, or a number in \code{[0,1]} that fixes pi0}

\item{control}{A list of control parameters  to be passed to the optimization function. `nlm` is  used.}
}
\value{
A list containing elements:
    \describe{
      \item{\code{posterior}}{A data frame of summary results (posterior
        means, and to add posterior log mean).}
      \item{\code{fitted_g}}{The fitted prior \eqn{\hat{g}} of class \code{point_gamma}} 
      \item{\code{log_likelihood}}{The optimal log likelihood attained
        \eqn{L(\hat{g})}.}
     }
}
\description{
Uses Empirical Bayes to fit the model \deqn{x_j | \lambda_j ~ Poi(s_j \lambda_j)} with \deqn{lambda_j ~ g()}
with Point Gamma: g()  = pi_0 delta() + (1-pi_0) gamma(shape, scale)
}
\details{
The model is fit in two stages: i) estimate \eqn{g} by maximum likelihood (over pi_0, shape, scale)
ii) Compute posterior distributions for \eqn{\lambda_j} given \eqn{x_j,\hat{g}}.
}
\examples{
   beta = c(rep(0,50),rexp(50))
   x = rpois(100,beta) # simulate Poisson observations
   s = replicate(100,1)
   out = ebpm_point_gamma(x,s)
}
